+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release Notes for Sprig Core
 
+## 2.1.0 - Unreleased
+### Added
+- Added the [s-replace-url](https://putyourlightson.com/plugins/sprig#s-replace-url) attribute that allows you to replace the current url of the browser location history.
+- Added the [s-select-oob](https://putyourlightson.com/plugins/sprig#s-select-oob) attribute that selects one or more elements from a server response to swap in via an out-of-band swap.
+- Added the [sprig.location()](https://putyourlightson.com/plugins/sprig#sprig.location) template variable that triggers a client-side redirect without reloading the page.
+- Added the [sprig.replaceUrl()](https://putyourlightson.com/plugins/sprig#sprig.replaceurl) template variable that replaces the current URL in the location bar.
+- Added the [sprig.reswap()](https://putyourlightson.com/plugins/sprig#sprig.reswap) template variable that allows you to change the swap behaviour.
+
+### Changed
+- Updated htmx to version 1.8.0 ([release notes](https://htmx.org/posts/2022-07-12-htmx-1.8.0-is-released/)).
+
+### Fixed
+- Fixed a bug in the `sprig.pushUrl()` template variable.
+- 
 ## 2.0.6 - 2022-09-25
 ### Fixed
 - Fixed a bug in which tags containing line breaks were not being parsed ([#264](https://github.com/putyourlightson/craft-sprig/issues/264)).